requirements.txt
pylibscrypt
requests
websockets
pillow 
pytesseract
# sudo apt-get install tesseract-ocr
opencv-python